Remove namespace from services
